# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * RentAuto API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Vehicle {
  id: number;
  name: string;
  brand: string;
  category: string;
  transmission: string;
  fuel: string;
  seats: number;
  pricePerDay: number;
  status: string;
  rating: number;
  imageUrl: string;
  featured: boolean;
  /** @nullable */
  description?: string | null;
}

export interface VehicleInput {
  name: string;
  brand: string;
  category: string;
  transmission: string;
  fuel: string;
  seats: number;
  pricePerDay: number;
  status?: string;
  imageUrl: string;
  featured?: boolean;
  description?: string;
}

export interface VehicleUpdate {
  name?: string;
  brand?: string;
  category?: string;
  transmission?: string;
  fuel?: string;
  seats?: number;
  pricePerDay?: number;
  status?: string;
  imageUrl?: string;
  featured?: boolean;
  description?: string;
}

export interface Reservation {
  id: number;
  vehicleId: number;
  vehicleName: string;
  /** @nullable */
  vehicleImageUrl?: string | null;
  clientName: string;
  clientEmail: string;
  pickupDate: string;
  dropoffDate: string;
  totalPrice: number;
  status: string;
  createdAt: string;
}

export interface ReservationInput {
  vehicleId: number;
  clientName: string;
  clientEmail: string;
  pickupDate: string;
  dropoffDate: string;
}

export interface DashboardStats {
  totalVehicles: number;
  activeReservations: number;
  monthlyRevenue: number;
  newClients: number;
  availableVehicles: number;
  rentedVehicles: number;
}

export type ListVehiclesParams = {
category?: string;
transmission?: string;
fuel?: string;
maxPrice?: number;
featured?: boolean;
};