# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: RentAuto API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: vehicles
    description: Vehicle operations
  - name: reservations
    description: Reservation operations
  - name: stats
    description: Dashboard stats
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /vehicles:
    get:
      operationId: listVehicles
      tags: [vehicles]
      summary: List all vehicles
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: transmission
          in: query
          schema:
            type: string
        - name: fuel
          in: query
          schema:
            type: string
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: featured
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of vehicles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vehicle"
    post:
      operationId: createVehicle
      tags: [vehicles]
      summary: Create a new vehicle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehicleInput"
      responses:
        "201":
          description: Created vehicle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"

  /vehicles/{id}:
    get:
      operationId: getVehicle
      tags: [vehicles]
      summary: Get a vehicle by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vehicle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
        "404":
          description: Not found
    patch:
      operationId: updateVehicle
      tags: [vehicles]
      summary: Update a vehicle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehicleUpdate"
      responses:
        "200":
          description: Updated vehicle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
    delete:
      operationId: deleteVehicle
      tags: [vehicles]
      summary: Delete a vehicle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /reservations:
    get:
      operationId: listReservations
      tags: [reservations]
      summary: List all reservations
      responses:
        "200":
          description: List of reservations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reservation"
    post:
      operationId: createReservation
      tags: [reservations]
      summary: Create a reservation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReservationInput"
      responses:
        "201":
          description: Created reservation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"

  /reservations/{id}:
    delete:
      operationId: deleteReservation
      tags: [reservations]
      summary: Cancel a reservation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /stats:
    get:
      operationId: getStats
      tags: [stats]
      summary: Get dashboard stats
      responses:
        "200":
          description: Stats summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Vehicle:
      type: object
      required:
        - id
        - name
        - brand
        - category
        - transmission
        - fuel
        - seats
        - pricePerDay
        - status
        - rating
        - imageUrl
        - featured
      properties:
        id:
          type: integer
        name:
          type: string
        brand:
          type: string
        category:
          type: string
        transmission:
          type: string
        fuel:
          type: string
        seats:
          type: integer
        pricePerDay:
          type: number
        status:
          type: string
        rating:
          type: number
        imageUrl:
          type: string
        featured:
          type: boolean
        description:
          type: ["string", "null"]

    VehicleInput:
      type: object
      required:
        - name
        - brand
        - category
        - transmission
        - fuel
        - seats
        - pricePerDay
        - imageUrl
      properties:
        name:
          type: string
        brand:
          type: string
        category:
          type: string
        transmission:
          type: string
        fuel:
          type: string
        seats:
          type: integer
        pricePerDay:
          type: number
        status:
          type: string
        imageUrl:
          type: string
        featured:
          type: boolean
        description:
          type: string

    VehicleUpdate:
      type: object
      properties:
        name:
          type: string
        brand:
          type: string
        category:
          type: string
        transmission:
          type: string
        fuel:
          type: string
        seats:
          type: integer
        pricePerDay:
          type: number
        status:
          type: string
        imageUrl:
          type: string
        featured:
          type: boolean
        description:
          type: string

    Reservation:
      type: object
      required:
        - id
        - vehicleId
        - vehicleName
        - clientName
        - clientEmail
        - pickupDate
        - dropoffDate
        - totalPrice
        - status
        - createdAt
      properties:
        id:
          type: integer
        vehicleId:
          type: integer
        vehicleName:
          type: string
        vehicleImageUrl:
          type: ["string", "null"]
        clientName:
          type: string
        clientEmail:
          type: string
        pickupDate:
          type: string
        dropoffDate:
          type: string
        totalPrice:
          type: number
        status:
          type: string
        createdAt:
          type: string

    ReservationInput:
      type: object
      required:
        - vehicleId
        - clientName
        - clientEmail
        - pickupDate
        - dropoffDate
      properties:
        vehicleId:
          type: integer
        clientName:
          type: string
        clientEmail:
          type: string
        pickupDate:
          type: string
        dropoffDate:
          type: string

    DashboardStats:
      type: object
      required:
        - totalVehicles
        - activeReservations
        - monthlyRevenue
        - newClients
        - availableVehicles
        - rentedVehicles
      properties:
        totalVehicles:
          type: integer
        activeReservations:
          type: integer
        monthlyRevenue:
          type: number
        newClients:
          type: integer
        availableVehicles:
          type: integer
        rentedVehicles:
          type: integer