# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * RentAuto API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardStats {
  totalVehicles: number;
  activeReservations: number;
  monthlyRevenue: number;
  newClients: number;
  availableVehicles: number;
  rentedVehicles: number;
}