# lib/api-zod/src/generated/types/reservation.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * RentAuto API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Reservation {
  id: number;
  vehicleId: number;
  vehicleName: string;
  /** @nullable */
  vehicleImageUrl?: string | null;
  clientName: string;
  clientEmail: string;
  pickupDate: string;
  dropoffDate: string;
  totalPrice: number;
  status: string;
  createdAt: string;
}