# lib/api-zod/src/generated/types/reservationInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * RentAuto API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReservationInput {
  vehicleId: number;
  clientName: string;
  clientEmail: string;
  pickupDate: string;
  dropoffDate: string;
}