# lib/api-zod/src/generated/types/vehicle.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * RentAuto API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Vehicle {
  id: number;
  name: string;
  brand: string;
  category: string;
  transmission: string;
  fuel: string;
  seats: number;
  pricePerDay: number;
  status: string;
  rating: number;
  imageUrl: string;
  featured: boolean;
  /** @nullable */
  description?: string | null;
}