# lib/api-zod/src/generated/types/vehicleInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * RentAuto API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VehicleInput {
  name: string;
  brand: string;
  category: string;
  transmission: string;
  fuel: string;
  seats: number;
  pricePerDay: number;
  status?: string;
  imageUrl: string;
  featured?: boolean;
  description?: string;
}